+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/crypto/store/db/model/KeyInfoEntity.kt
@@ -29,4 +29,10 @@ internal open class KeyInfoEntity(
          */
         var signatures: String? = null,
         var trustLevelEntity: TrustLevelEntity? = null
-) : RealmObject()
+) : RealmObject() {
+
+    fun deleteOnCascade() {
+        trustLevelEntity?.deleteFromRealm()
+        deleteFromRealm()
+    }
+}